chore: refactor relay
