+++ package.json
@@ -42,6 +42,7 @@
     "devDependencies": {
         "@babel/eslint-parser": "^7.19.1",
         "@babel/preset-typescript": "^7.18.6",
+        "@types/chai": "^4.3.4",
         "@types/fs-extra": "^9.0.13",
         "@types/jest": "^29.2.1",
         "@types/lodash": "^4.14.186",
