+++ src/__tests__/test.js
@@ -20,222 +20,13 @@ const {
     forkNetwork,
     mainnetInfo,
     networks,
+    contracts: { BurnableMintableCappedERC20 },
 } = require('../');
-const BurnableMintableCappedERC20 = require('../../dist/artifacts/@axelar-network/axelar-cgp-solidity/contracts/BurnableMintableCappedERC20.sol/BurnableMintableCappedERC20.json');
 const { keccak256, toUtf8Bytes } = require('ethers/lib/utils');
 
-setLogger((...args) => {});
+setLogger(() => undefined);
 
 jest.setTimeout(300000);
-
-
-
-describe('relay', () => {
-    let chain1, chain2;
-    let user1, user2;
-    beforeEach(async () => {
-        chain1 = await createNetwork({ seed: 1 });
-        [user1] = chain1.userWallets;
-        chain1.usdc = await chain1.deployToken('Axelar Wrapped USDC', 'aUSDC', 6, 0);
-        chain2 = await createNetwork({ seed: 2 });
-        [user2] = chain2.userWallets;
-        chain2.usdc = await chain2.deployToken('Axelar Wrapped USDC', 'aUSDC', 6, 0);
-    });
-    afterEach(async () => {
-        stopAll();
-    });
-    describe('deposit address', () => {
-        it('should generate a deposit address', async () => {
-            const depositAddress = getDepositAddress(chain1, chain2, user2.address, 'aUSDC');
-            const amount = BigInt(12423532412);
-            const fee = BigInt(getFee(chain1, chain2, 'aUSDC'));
-            await chain1.giveToken(user1.address, 'aUSDC', amount);
-            await (await chain1.usdc.connect(user1).transfer(depositAddress, amount)).wait();
-            await relay();
-
-            expect(BigInt(await chain2.usdc.balanceOf(user2.address))).to.equal(amount - fee);
-        });
-        it('should generate a deposit address to use twice', async () => {
-            const depositAddress = getDepositAddress(chain1, chain2, user2.address, 'aUSDC');
-            const amount1 = BigInt(12423532412);
-            const amount2 = BigInt(5489763092348);
-            const fee = BigInt(getFee(chain1, chain2, 'aUSDC'));
-            await chain1.giveToken(user1.address, 'aUSDC', amount1);
-            await (await chain1.usdc.connect(user1).transfer(depositAddress, amount1)).wait();
-            await relay();
-            expect(BigInt(await chain2.usdc.balanceOf(user2.address))).to.equal(amount1 - fee);
-
-            await chain1.giveToken(user1.address, 'aUSDC', amount2);
-            await (await chain1.usdc.connect(user1).transfer(depositAddress, amount2)).wait();
-            await relay();
-            expect(BigInt(await chain2.usdc.balanceOf(user2.address))).to.equal(amount1 - fee + amount2 - fee);
-        });
-        it('should generate a deposit address remotely', async () => {
-            const port = 8501;
-            await new Promise((resolve) => {
-                listen(port, resolve());
-            });
-            const depositAddress = await getDepositAddress(chain1, chain2, user2.address, 'aUSDC', port);
-            const amount = BigInt(12423532412);
-            const fee = BigInt(getFee(chain1, chain2, 'aUSDC'));
-            await chain1.giveToken(user1.address, 'aUSDC', amount);
-            await (await chain1.usdc.connect(user1).transfer(depositAddress, amount)).wait();
-            await relay();
-            expect(BigInt(await chain2.usdc.balanceOf(user2.address))).to.equal(amount - fee);
-        });
-    });
-    describe('send token', () => {
-        it('should send some usdc over', async () => {
-            const amount = BigInt(1e8);
-            const fee = BigInt(getFee(chain1, chain2, 'aUSDC'));
-            await chain1.giveToken(user1.address, 'aUSDC', amount);
-            await (await chain1.usdc.connect(user1).approve(chain1.gateway.address, amount)).wait();
-            await (await chain1.gateway.connect(user1).sendToken(chain2.name, user2.address, 'aUSDC', amount)).wait();
-            await relay();
-            expect(BigInt(await chain2.usdc.balanceOf(user2.address))).to.equal(amount - fee);
-        });
-    });
-    describe('call contract', () => {
-        let ex1, ex2;
-        const Executable = require('../artifacts/src/contracts/test/Executable.sol/Executable.json');
-
-        const message = 'hello there executables!';
-        const payload = defaultAbiCoder.encode(['string'], [message]);
-        beforeEach(async () => {
-            ex1 = await deployContract(user1, Executable, [chain1.gateway.address, chain1.gasService.address]);
-            ex2 = await deployContract(user2, Executable, [chain2.gateway.address, chain1.gasService.address]);
-
-            await await ex1.connect(user1).addSibling(chain2.name, ex2.address);
-            await await ex2.connect(user2).addSibling(chain1.name, ex1.address);
-        });
-        it('should call a contract manually and fulfill the call', async () => {
-            await (await chain1.gateway.connect(user1).callContract(chain2.name, ex2.address, payload)).wait();
-            await relay();
-            const filter = chain2.gateway.filters.ContractCallApproved();
-            const args = (await chain2.gateway.queryFilter(filter))[0].args;
-            await (await ex2.connect(user2).execute(args.commandId, chain1.name, user1.address, payload)).wait();
-
-            expect(await ex1.value()).to.equal('');
-            expect(await ex2.value()).to.equal(message);
-            expect(await ex2.sourceChain()).to.equal(chain1.name);
-            expect(await ex2.sourceAddress()).to.equal(user1.address);
-        });
-        it('should pay for gas and call a contract manually', async () => {
-            await (
-                await chain1.gasService
-                    .connect(user1)
-                    .payNativeGasForContractCall(user1.address, chain2.name, ex2.address, payload, user1.address, { value: 1e6 })
-            ).wait();
-            await (await chain1.gateway.connect(user1).callContract(chain2.name, ex2.address, payload)).wait();
-            await relay();
-
-            expect(await ex1.value()).to.equal('');
-            expect(await ex2.value()).to.equal(message);
-            expect(await ex2.sourceChain()).to.equal(chain1.name);
-            expect(await ex2.sourceAddress()).to.equal(user1.address);
-        });
-        it('should call a contract through the sibling and fulfill the call', async () => {
-            await (await ex1.connect(user1).set(chain2.name, message)).wait();
-            await relay();
-            const filter = chain2.gateway.filters.ContractCallApproved();
-            const args = (await chain2.gateway.queryFilter(filter))[0].args;
-            await (await ex2.connect(user2).execute(args.commandId, chain1.name, ex1.address, payload)).wait();
-
-            expect(await ex1.value()).to.equal(message);
-            expect(await ex2.value()).to.equal(message);
-            expect(await ex2.sourceChain()).to.equal(chain1.name);
-            expect(await ex2.sourceAddress()).to.equal(ex1.address);
-        });
-        it('should have the sibling pay for gas and make the call', async () => {
-            await (await ex1.connect(user1).set(chain2.name, message, { value: BigInt(1e18) })).wait();
-            await relay();
-
-            expect(await ex1.value()).to.equal(message);
-            expect(await ex2.value()).to.equal(message);
-            expect(await ex2.sourceChain()).to.equal(chain1.name);
-            expect(await ex2.sourceAddress()).to.equal(ex1.address);
-        });
-    });
-    describe('call contract with token', () => {
-        let ex1, ex2;
-        const Executable = require('../artifacts/src/contracts/test/ExecutableWithToken.sol/ExecutableWithToken.json');
-
-        const message = 'hello there executables!';
-        const amount = 1234255675;
-        let payload;
-
-        beforeEach(async () => {
-            payload = defaultAbiCoder.encode(['string', 'address'], [message, user2.address]);
-            ex1 = await deployContract(user1, Executable, [chain1.gateway.address, chain1.gasService.address]);
-            ex2 = await deployContract(user2, Executable, [chain2.gateway.address, chain1.gasService.address]);
-
-            await await ex1.connect(user1).addSibling(chain2.name, ex2.address);
-            await await ex2.connect(user2).addSibling(chain1.name, ex1.address);
-
-            await chain1.giveToken(user1.address, 'aUSDC', amount);
-        });
-        it('should call a contract manually and fulfill the call', async () => {
-            await (await chain1.usdc.connect(user1).approve(chain1.gateway.address, amount)).wait();
-            await (await chain1.gateway.connect(user1).callContractWithToken(chain2.name, ex2.address, payload, 'aUSDC', amount)).wait();
-            await relay();
-            const filter = chain2.gateway.filters.ContractCallApprovedWithMint();
-            const args = (await chain2.gateway.queryFilter(filter))[0].args;
-            await (await ex2.connect(user2).executeWithToken(args.commandId, chain1.name, user1.address, payload, 'aUSDC', amount)).wait();
-
-            expect(await ex1.value()).to.equal('');
-            expect(await ex2.value()).to.equal(message);
-            expect(await ex2.sourceChain()).to.equal(chain1.name);
-            expect(await ex2.sourceAddress()).to.equal(user1.address);
-            expect((await chain2.usdc.balanceOf(user2.address)).toNumber()).to.equal(amount);
-        });
-        it('should pay for gas and call a contract manually', async () => {
-            await await chain1.gasService
-                .connect(user1)
-                .payNativeGasForContractCallWithToken(user1.address, chain2.name, ex2.address, payload, 'aUSDC', amount, user1.address, {
-                    value: 1e6,
-                });
-
-            await (await chain1.usdc.connect(user1).approve(chain1.gateway.address, amount)).wait();
-            await (await chain1.gateway.connect(user1).callContractWithToken(chain2.name, ex2.address, payload, 'aUSDC', amount)).wait();
-            await relay();
-
-            expect(await ex1.value()).to.equal('');
-            expect(await ex2.value()).to.equal(message);
-            expect(await ex2.sourceChain()).to.equal(chain1.name);
-            expect(await ex2.sourceAddress()).to.equal(user1.address);
-            expect((await chain2.usdc.balanceOf(user2.address)).toNumber()).to.equal(amount);
-        });
-        it('should call a contract through the sibling and fulfill the call', async () => {
-            await (await chain1.usdc.connect(user1).approve(ex1.address, amount)).wait();
-            await (await ex1.connect(user1).setAndSend(chain2.name, message, user2.address, 'aUSDC', amount)).wait();
-            await relay();
-            const filter = chain2.gateway.filters.ContractCallApprovedWithMint();
-            const args = (await chain2.gateway.queryFilter(filter))[0].args;
-            await (await ex2.connect(user2).executeWithToken(args.commandId, chain1.name, ex1.address, payload, 'aUSDC', amount)).wait();
-
-            expect(await ex1.value()).to.equal(message);
-            expect(await ex2.value()).to.equal(message);
-            expect(await ex2.sourceChain()).to.equal(chain1.name);
-            expect(await ex2.sourceAddress()).to.equal(ex1.address);
-            expect((await chain2.usdc.balanceOf(user2.address)).toNumber()).to.equal(amount);
-        });
-        it('should have the sibling pay for gas and make the call', async () => {
-            await (await chain1.usdc.connect(user1).approve(ex1.address, amount)).wait();
-            await (
-                await ex1.connect(user1).setAndSend(chain2.name, message, user2.address, 'aUSDC', amount, {
-                    value: 1e6,
-                })
-            ).wait();
-            await relay();
-
-            expect(await ex1.value()).to.equal(message);
-            expect(await ex2.value()).to.equal(message);
-            expect(await ex2.sourceChain()).to.equal(chain1.name);
-            expect(await ex2.sourceAddress()).to.equal(ex1.address);
-            expect((await chain2.usdc.balanceOf(user2.address)).toNumber()).to.equal(amount);
-        });
-    });
-});
 
 describe.skip('forking', () => {
     afterEach(async () => {
